fzf does not close
